Feature/article

+++ controllers/articleController.js
@@ -5,8 +5,14 @@ import {
   getArticleById,
   updateArticle,
 } from "../models/articleModel.js";
+import { articleFormSchema } from "../validation/article/articleValidation.js";
 
 export const create = async (req, res) => {
+  const { error } = articleFormSchema.validate(req.body);
+  if (error) {
+    return res.status(400).json({ error: error.message });
+  }
+
   const { title, content, status } = req.body;
   const image = req.file ? req.file.filename : null;
 
@@ -55,11 +61,12 @@ export const getAll = async (req, res) => {
 
 export const getOne = async (req, res) => {
   try {
+    console.log(req.params.id);
     const article = await getArticleById(req.params.id);
     if (!article) {
-      return res.status(200).json({
-        status: "Succes",
-        code: 200,
+      return res.status(404).json({
+        status: "Error",
+        code: 404,
         message: "Article not found",
         data: [],
       });
@@ -82,22 +89,49 @@ export const getOne = async (req, res) => {
 };
 
 export const update = async (req, res) => {
+  const { error } = articleFormSchema.validate(req.body);
+  if (error) {
+    return res.status(400).json({ error: error.message });
+  }
+
   const { title, content, status } = req.body;
-  const image = req.file ? req.file.filename : null;
 
   try {
-    const result = await updateArticle(
+    const article = await getArticleById(req.params.id);
+
+    if (!article) {
+      return res.status(404).json({
+        status: "Error",
+        code: 404,
+        message: "Article not found",
+        data: [],
+      });
+    }
+
+    if (article.user_id !== req.user.id) {
+      return res.status(403).json({
+        status: "Error",
+        code: 403,
+        message: "You are not allowed to update this article",
+        data: [],
+      });
+    }
+
+    const image = req.file ? req.file.filename : article.image;
+
+    const updatedArticle = await updateArticle(
       req.params.id,
       title,
       content,
       image,
       status
     );
+
     res.status(200).json({
       status: "Success",
       code: 200,
       message: "Success update article",
-      data: result,
+      data: updatedArticle,
     });
   } catch (err) {
     res.status(500).json({
@@ -111,7 +145,27 @@ export const update = async (req, res) => {
 
 export const remove = async (req, res) => {
   try {
+    const article = await getArticleById(req.params.id);
+    if (!article) {
+      return res.status(404).json({
+        status: "Error",
+        code: 404,
+        message: "Article not found",
+        data: [],
+      });
+    }
+
+    if (article.user_id !== req.user.id) {
+      return res.status(403).json({
+        status: "Error",
+        code: 403,
+        message: "You are not allowed to delete this article",
+        data: [],
+      });
+    }
+
     await deleteArticle(req.params.id);
+
     res.status(200).json({
       status: "Success",
       code: 200,
